requirements.txt
# Bot dependencies
python-telegram-bot==20.7
aiohttp==3.9.3
python-dotenv==1.0.1
nest-asyncio==1.6.0

# Web App dependencies
Flask==3.0.2
flask_sqlalchemy==3.1.1
flask_cors==4.0.0
mysqlclient==2.2.4
sqlalchemy-utils==0.41.1
gunicorn==21.2.0
